lexichat-api/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.3.0+cpu
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.22
python-dotenv==1.2.2
sqlalchemy==2.0.48
passlib==1.7.4
bcrypt==5.0.0
python-jose==3.5.0
pydantic==2.12.5
pydantic[email]==2.12.5
email-validator==2.3.0
pinecone==8.1.0
groq==1.1.1
sentence-transformers==3.4.1
requests
PyMuPDF==1.25.3
python-docx==1.2.0
psycopg2-binary==2.9.9
